+++ packages/util-crypto/src/address/sort.ts
@@ -1,6 +1,7 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { HexString } from '@polkadot/util/types';
 import type { Prefix } from './types';
 
 import { u8aSorted } from '@polkadot/util';
@@ -8,7 +9,7 @@ import { u8aSorted } from '@polkadot/util';
 import { decodeAddress } from './decode';
 import { encodeAddress } from './encode';
 
-export function sortAddresses (addresses: (Uint8Array | string)[], ss58Format?: Prefix): string[] {
+export function sortAddresses (addresses: (HexString | Uint8Array | string)[], ss58Format?: Prefix): string[] {
   return u8aSorted(
     addresses.map((who) => decodeAddress(who))
   ).map((u8a) => encodeAddress(u8a, ss58Format));