+++ packages/util-crypto/src/json/types.ts
@@ -1,6 +1,8 @@
 // Copyright 2017-2021 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { HexString } from '@polkadot/util/types';
+
 export type EncryptedJsonVersion = '0' | '1' | '2' | '3';
 
 export type EncryptedJsonEncoding = 'none' | 'scrypt' | 'xsalsa20-poly1305';
@@ -12,6 +14,6 @@ export interface EncryptedJsonDescriptor {
 }
 
 export interface EncryptedJson {
-  encoded: string;
+  encoded: HexString | string;
   encoding: EncryptedJsonDescriptor;
 }