+++ packages/util-crypto/src/keccak/asHex.ts
@@ -1,6 +1,8 @@
 // Copyright 2017-2021 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { HexString } from '@polkadot/util/types';
+
 import { u8aToHex } from '@polkadot/util';
 
 import { keccakAsU8a } from './asU8a';
@@ -19,7 +21,7 @@ import { keccakAsU8a } from './asU8a';
  * keccakAsHex('123'); // => 0x...
  * ```
  */
-export function keccakAsHex (value: Buffer | Uint8Array | string, bitLength?: 256 | 512): string {
+export function keccakAsHex (value: HexString | Buffer | Uint8Array | string, bitLength?: 256 | 512): HexString {
   return u8aToHex(
     keccakAsU8a(value, bitLength)
   );