+++ packages/util-crypto/src/schnorrkel/keypair/fromSeed.ts
@@ -1,6 +1,7 @@
 // Copyright 2017-2021 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { HexString } from '@polkadot/util/types';
 import type { Keypair } from '../../types';
 
 import { assert, u8aToU8a } from '@polkadot/util';
@@ -12,7 +13,7 @@ import { schnorrkelKeypairFromU8a } from './fromU8a';
  * @name schnorrkelKeypairFromSeed
  * @description Returns a object containing a `publicKey` & `secretKey` generated from the supplied seed.
  */
-export function schnorrkelKeypairFromSeed (seed: Uint8Array | string): Keypair {
+export function schnorrkelKeypairFromSeed (seed: HexString | Uint8Array | string): Keypair {
   const seedU8a = u8aToU8a(seed);
 
   assert(seedU8a.length === 32, () => `Expected a seed matching 32 bytes, found ${seedU8a.length}`);